+++ src/main/java/ua/goit/service/SumOfGood.java
@@ -9,7 +9,7 @@ public class SumOfGood implements SumOfGoodIntf<String>{
     private final BaseIntf<String, Good> goods = Factory.of(Good.class);
 
     @Override
-    public Double calculateSumOfGood(String id, Integer quantity) {
+    public Double calculateSumOfGood(String id, Long quantity) {
         Integer promoQuantity = goods.getOne(id).getPromo().getPromoQuantity();
         return (promoQuantity == 0 ) ?
                 goods.getOne(id).getPrice() * quantity :
