+++ src/main/java/ua/goit/service/SumOfGoodIntf.java
@@ -2,7 +2,7 @@ package ua.goit.service;
 
 public interface SumOfGoodIntf <ID>{
 
-    Double calculateSumOfGood(ID id, Integer quantity);
+    Double calculateSumOfGood(ID id, Long quantity);
 
     static SumOfGoodIntf of() {
         return new SumOfGood();
